kill kill kill them all kill them till they're dead merrily merrily merrily merrily shoot them in the head.
oh wait
they're already dead and I already wrote this fucking script
and they're still mimicking assholes from the 80s and 90s that are dead
what fun !

and I already wrote this

and a cop pretendd to fucking care
and i really want a gun right now.
but as we all know the cops aren't real at this point

and i really really wish i could be proactive and shoot them all.
jesus.